requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
os-sys
pillow
pybase64
Python-IO

